Creating edit and delete profile routes and views

+++ routes/profile.routes.js
@@ -5,7 +5,7 @@ const router = express.Router();
 
 const User = require('../models/User.model')
 const Gif = require('../models/Gif.model');
-const { exists } = require('../models/User.model');
+const { readSync } = require('node:fs');
 
 // ----------MIDDLEWARE---------//
 
@@ -26,7 +26,7 @@ router.get('/profile', checkForAuth ,(req, res) => {
     User.findById(req.user._id)
         .populate('favGifs')
         .then(result => {
-          res.render('profile/profile', {user: result, layout: layout, fav: anyFavorites})
+          res.status(200).render('profile/profile', {user: result, layout: layout, fav: anyFavorites})
         })
         .catch(error => {
           console.log(error)
@@ -76,21 +76,47 @@ router.post('/add-gif', checkForAuth, (req, res)=>{
 router.post('/remove-gif/:_id', checkForAuth, (req, res) => {
   User.findByIdAndUpdate(req.user._id, {$pull: {favGifs: req.params._id}})
     .then(() => {
-      res.redirect('/profile')
+      res.status(200).redirect('/profile')
     })
     .catch(error => {
       console.log(error)
     })
 })
 
+router.get('/edit-user', checkForAuth, (req, res)=>{
+  const layout = req.user ? '/layouts/auth' : '/layouts/noAuth'
+  res.status(200).render('profile/edit-profile', {layout: layout})
+})
+
+router.post('/edit-user', checkForAuth, (req, res)=>{
+  User.findByIdAndUpdate(req.user._id, req.body)
+    .then(result => {
+      res.status(200).redirect('/profile')
+    })
+    .catch(error => {
+      console.log(error)
+    })
+    
+})
+
+router.get('/delete-account', checkForAuth, (req, res)=>{
+  const layout = req.user ? '/layouts/auth' : '/layouts/noAuth'
+  res.render('profile/delete-profile', {user: req.user, layout: layout})
+})
+
 router.post('/delete-account/', checkForAuth, (req, res)=>{
+  if (user.req.username === req.body.username){
     User.findByIdAndDelete(req.user._id)
     .then(() => {
-        res.redirect('/')
+        res.status(200).redirect('/')
     })
     .catch(error=>{
         console.log(error)
     })
+  }else{
+    const layout = req.user ? '/layouts/auth' : '/layouts/noAuth'
+    res.redirect('/delete-profile', {user: req.user, layout: layout})
+  }
 })
 
 module.exports = router
